Backend - Bug fix of validating id from query

+++ backend/src/controllers/reserve.controller.js
@@ -1,19 +1,27 @@
 /* eslint-disable no-underscore-dangle */
+import bikesService from '../services/bikes.service';
 import reserveService from '../services/reserves.service';
 import {MANAGER} from '../utils/constants';
+import isValidMongooseObjectId from '../utils/mongoose';
 
 class ReserveController {
   async listReserves(req, res) {
     try {
       const page = parseInt(req.query.page, 10) || 0;
       const {user} = req.query;
+      const {bike} = req.query;
+      const isValidUserId = isValidMongooseObjectId(user);
+      const isValidBikeId = isValidMongooseObjectId(bike);
+      if ((user && !isValidUserId) || (bike && !isValidBikeId)) {
+        return res.status(400).json({success: false, error: 'Not valid id'});
+      }
       if (req.user.role !== MANAGER && user !== req.user.id) {
         return res.status(403).json({
           success: false,
           error: 'Not authorised.',
         });
       }
-      const reserves = await reserveService.list(user, page, 10);
+      const reserves = await reserveService.list(user, bike, page, 10);
       res.status(200).json({success: true, data: reserves});
     } catch (e) {
       console.error(e);
@@ -24,10 +32,7 @@ class ReserveController {
   async getReserveById(req, res) {
     try {
       const reserve = await reserveService.findById(req.params.reserveId);
-      if (
-        req.user.role !== MANAGER &&
-        reserve.user._id.toString() !== req.user.id
-      ) {
+      if (reserve.user._id.toString() !== req.user.id) {
         return res.status(403).json({
           success: false,
           error: 'Not authorised.',
@@ -54,6 +59,12 @@ class ReserveController {
           success: false,
           error: 'Not authorised.',
         });
+      }
+      const isAvailable = (await bikesService.findById(bike)).available;
+      if (!isAvailable) {
+        return res
+          .status(400)
+          .json({success: false, error: 'Bike not available'});
       }
       const isReserved = await reserveService.isBikeReserved(bike, from, to);
       console.log(isReserved);
@@ -102,7 +113,14 @@ class ReserveController {
           .status(404)
           .json({success: false, error: 'Reserve not found'});
       }
-
+      const isReserveCompleted = await reserveService.isReservedCompleted(
+        req.params.reserveId,
+      );
+      if (isReserveCompleted) {
+        return res
+          .status(400)
+          .json({success: false, error: 'Reserve period already over'});
+      }
       await reserveService.deleteById(req.params.reserveId);
       res.status(204).json({});
     } catch (e) {
